+++ lib/src/themes/colors.dart
@@ -1,0 +1,6 @@
+import 'dart:ui';
+
+class DefaultColors {
+  static const messageSelf = Color.fromARGB(255, 58, 136, 102);
+  static const messageOther = Color.fromARGB(233, 68, 137, 255);
+}